requirements.txt
tqdm
numpy
# torch<2.0.0
# torchvision
omegaconf
matplotlib
Pillow>=9.2.0
huggingface_hub>=0.13.2
transformers~=4.25.1
accelerate~=0.15.0
diffusers~=0.16.0
tokenizers~=0.13.2
sentencepiece~=0.1.97
ftfy~=6.1.1
beautifulsoup4~=4.11.1
